+++ CHANGELOG.md
@@ -5,17 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [r11.6.1] - 2021-05-17
+## [r11.6.1] - 2021-05-18
 
 ### Added
 - Image component with custom render if loading error.
 - ScreenWrapper take sceneBackground as backgroundColor.
 - Common component for social like Feeds, ActionBar.
 - Like, comment, share for news.
-
 ### Changed
-- Container can select wrapper between reanimated and animated.
+- Container can select reanimated or animated wrapper.
 - UI/ UX, logic of News (Notify) screen.
+
 ## [r11.5.5] - 2021-05-08
 ### Fixed
 - Fix resend otp calling.